Identify and bandaid issue
The underlying issue was that the SphericalMercator package was
defaulting to creating WGS84 bounding boxes which worked "fine"
on our WGS84 data but led to an incorrectly projected map. Mapnik
doesn't actually autoconvert projections (despite being able to
specify a source/output srs) so data was being drawn as WGS84.
Specifying '900913' in the bounding box method and manually converting
the database data to web mercator solved the misalignment.