Remove unhelpful vector error handling
The vector tile endpoint now no longer catches errors, because it's
more helpful for the client to recieve a 500 error than json or a
message tile. The entire error handling system should probably be
changed at some point to return errors that more accurately reflect
the one that's happening on the server.