:memo: Update readme in light of new deploy script
I forgot to include ./scripts/destroy in the docs, so this adds it
in in place of ./scripts/publish under the Deployment heading.
Instructions on setting up a config file have also been added to this
section for the benefit of those who just want to deploy rather than
set the project up for development.