Use pre-installed AWS SDK on Lambda
The Lambda environment comes with the AWS SDK pre-installed, so it wastes
space in the bundle to install it ourselves.

Per [this issue](https://github.com/claudiajs/claudia/issues/82), just moving
it to `devDependencies` won't work because Claudia validates the bundle and
it's an error for the package to be missing, but that can be worked around
by making at an optional dependency and telling Claudia not to install
optional.

Also bumps the version listed in package.json (and installed in development)
to the current version pre-installed on Lambda.

Resolves #118.