Add S3 cache between CloudFront proxy and API
Adds configuration for an S3 bucket, configured as a website, to serve
as a tile cache. Rather than sending requests right to the API, the
CloudFront distribution will now request them from the S3 endpoint, which
in turn will redirect to the API for tiles it doesn't have.

Still to do: add code to the Lambda function to write generated tiles
back to S3.