Switch API builder to the API Gateway Proxy request object
See https://claudiajs.com/news/2016/09/27/claudia-2.0.html#request

I was hoping it would make it easier to get at the original URL and
querystring, but it doesn't seem to. But it seems like the `req` object
we've been using is the "for backwards compatibility" one and the
AWS_PROXY one is potentially more robust and portable.