Write tiles to S3, ignoring querystrings
It turns out S3 websites use only the path, not the querystring, as the
key. Which makes sense, but creates a problem for our setup where some of
the parameters that define the tile are in the querystring.

This gets S3 caching working, but with querystrings ignored. We'll need
to refactor the API to change all query parameters to path parameters.