Add separate script for decrypting environment file
The 'deploy-demo.sh' script is run by the Travis 'deploy' step on the
branch being built, so it's not a great place for the command to decrypt
the encrypted .env file used for the demo deploy. Putting it in a separate
script means we can call it from within 'deploy-demo.sh' *after* checking
out the gh-pages branch, so it will pick up changes to the command (e.g.
if the encryption key changes).