Decrypt demo .env in a separate script
The version of 'deploy-demo.sh' on the branch being tested runs, not the
version on gh-pages, so having the command to decrypt the demo .env file
in that script means any change to the command (e.g. the key IDs) needs to
be merged to develop.  This puts it in a separate script so, when it gets
invoked after the gh-pages branch is checked out, it's the one on the
branch that runs.