Switch to compiling config XML in advance but filling vars at runtime
Changes how map config is handled:
- Makes the Mapnik XML canonical, and tracked.
- Moves the Carto MML/MSS map config files into a subdirectory. They are
now useful as working files, but aren't used by the running code.
- Removes the compilation step from the `deploy` and `local` yarn commands.
Config file compilation can still be run using the `build-xml` script, but
won't be run automatically.
- Removes the 'template-vars' function that was being used during XML
compilation, moving the logic into a function in tiler.js. The XML files
will now be templates to be filled with variables from the environment,
like the Carto map config files used to be, and the substitutions will
be made at runtime.
- Changes the variables and how they're provided to docker-compose to
support the new setup of filling them in at runtime. This includes adding
commands to the `deploy` scripts to add the variables to the Lambda
function's environment.