Set up docker-compose and deploy scripts to use env-specific env files
Sets up the docker-compose config and deployment scripts to take advantage
of the new (or impending) docker-compose `--env-file` option. Creating a
.env.ENVIRON file then passing the name of the environment to `scripts/deploy`
should read that environment file into the project environment.
All the variables needed inside the tiler container have to be listed in
it's `environment` so they get passed through.

Adds a checked-in .env.local file, with credentials for the test database.
It doesn't need any of the Lambda or project variables, since it's local.