Add an ENVIRON variable and use it to set deployment name
Changes the deployment-related scripts to expect an argument for which
environment to deploy. They'll look for a `.env.ENVIRON` file to set the
environment variables and will create claudia-ENVIRON.json and
.api-id-ENVIRON files rather than always using the same unqualified names.

Using separate claudia.json files means the config file is now given as
an argument, rather than using the default, and they go in a subdirectory.