Adjust Terraform for multi-env setup
Changes the terraform deployment steps to use local tfplan and tfstate
files to keep track of separate environments. The "local" backend is
the default, but specifying it explicitly enables specifying parameters
for it, i.e. the ENVIRON-based filenames to use.

This also switches from using `plan -auto-apply` to doing separate plan
and apply commands, though they're still within the same `publish` script.
The behavior will end up the same, but it now gives you 10 seconds to
bail out if the plan looks weird, rather than applying it as soon as it's
calculated.